% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{elem_boys}
\alias{elem_boys}
\title{Boy Baby names.}
\format{
A data.table with 2 variables: \code{types}, \code{counts},
}
\usage{
elem_boys
}
\description{
Full baby name data provided by the SSA. This includes all names with at
least 5 uses.
}
\keyword{datasets}
